+++ CHANGES.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [dev-3.1.0-1]
+- [WIP] Adjust Backend module statistic to multiple answers
+- [FEATURE] Calculation for quiz result adjusted to multiple correct answers
+
+## [dev-3.1.0-1]
 - [WIP] Calculation for quiz result adjusted to multiple correct answers
 - [FEATURE] Answering output need to conside multiple correct answers